cybersec_assistant_gemini_flash/CyberSec_Assistant/requirements.txt
streamlit>=1.28.0
google-generativeai>=0.5.0
yara-python>=4.3.1
requests>=2.31.0
python-nmap>=0.7.1